pyproject.toml
[tool.poetry]
name = "d-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.5.0"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
